+++ crates/core/src/tool.rs
@@ -210,8 +210,8 @@ impl Tool {
 
 impl Tool {
     /// Return contextual information to pass to WASM plugin functions.
-    pub fn create_context(&self) -> ToolContext {
-        ToolContext {
+    pub fn create_plugin_context(&self) -> PluginContext {
+        PluginContext {
             proto_version: Some(get_proto_version().to_owned()),
             temp_dir: self.to_virtual_path(self.get_temp_dir()),
             tool_dir: self.to_virtual_path(self.get_product_dir()),
@@ -222,8 +222,8 @@ impl Tool {
     /// Return contextual information to pass to WASM plugin functions,
     /// representing an unresolved state, which has no version or tool
     /// data.
-    pub fn create_unresolved_context(&self) -> ToolUnresolvedContext {
-        ToolUnresolvedContext {
+    pub fn create_plugin_unresolved_context(&self) -> PluginUnresolvedContext {
+        PluginUnresolvedContext {
             proto_version: Some(get_proto_version().to_owned()),
             temp_dir: self.to_virtual_path(&self.inventory.temp_dir),
             // version: self.version.clone(),