+++ CHANGELOG.md
@@ -15,6 +15,12 @@
 - [Rust](https://github.com/moonrepo/plugins/blob/master/tools/rust/CHANGELOG.md)
 - [Schema (TOML, JSON, YAML)](https://github.com/moonrepo/plugins/blob/master/tools/internal-schema/CHANGELOG.md)
 
+## Unreleased
+
+- **WASM API**
+  - Renamed `ToolContext` to `PluginContext`.
+  - Renamed `ToolUnresolvedContext` to `PluginUnresolvedContext`.
+
 ## 0.51.6
 
 #### 🚀 Updates